1 Added better descriptions for standard colours 2. Moved QMColours class to src/gluon/util 3. Linting
