--- src/QMColour.ts
@@ -1,30 +1,0 @@
-export class QMColours {
-
-    public static absaAgile: QMColour = { name: "absaAgile", hex: "#f52d28", description: "absaAgile" };
-    public static absaBlack: QMColour = { name: "absaBlack", hex: "#000000", description: "absaBlack" };
-    public static absaCalm: QMColour = { name: "absaCalm", hex: "#870a3c", description: "absaCalm" };
-    public static absaCare: QMColour = { name: "absaCare", hex: "#f05a7d", description: "absaCare" };
-    public static absaDepth: QMColour = { name: "absaDepth", hex: "#500a28", description: "absaDepth" };
-    public static absaEnergy: QMColour = { name: "absaEnergy", hex: "#ff780f", description: "absaEnergy" };
-    public static absaGrounded: QMColour = { name: "absaGrounded", hex: "#960528", description: "absaGrounded" };
-    public static absaHuman: QMColour = { name: "absaHuman", hex: "#aa052d", description: "absaHuman" };
-    public static absaLuxury: QMColour = { name: "absaLuxury", hex: "#640032", description: "absaLuxury" };
-    public static absaPassion: QMColour = { name: "absaPassion", hex: "#dc0032", description: "absaPassion" };
-    public static absaPrepared: QMColour = { name: "absaPrepared", hex: "#fa551e", description: "absaPrepared" };
-    public static absaSmile: QMColour = { name: "absaSmile", hex: "#f0325a", description: "absaSmile" };
-    public static absaSurprise: QMColour = { name: "absaSurprise", hex: "#af144b", description: "absaSurprise" };
-    public static absaWarmth: QMColour = { name: "absaWarmth", hex: "#be0028", description: "absaWarmth" };
-
-    public static stdGreenyMcAppleStroodle: QMColour = { name: "stdGreenyMcAppleStroodle", hex: "#45B254", description: "stdGreenyMcAppleStroodle" };
-    public static stdMuddyYellow: QMColour = { name: "stdMuddyYellow", hex: "#ffcc00", description: "stdMuddyYellow" };
-    public static stdPurplePeopleEater: QMColour = { name: "stdPurplePeopleEater", hex: "#c000ff", description: "stdPurplePeopleEater" };
-    public static stdReddyMcRedFace: QMColour = { name: "stdReddyMcRedFace", hex: "#D94649", description: "stdReddyMcRedFace" };
-    public static stdShySkyBlue: QMColour = { name: "stdShySkyBlue", hex: "#00a5ff", description: "stdShySkyBlue" };
-    public static stdTurquoiseSurprise: QMColour = { name: "stdTurquoise Surprise", hex: "#00ddff", description: "stdTurquoise Surprise" };
- }
-
-export class QMColour {
-    public name: string;
-    public hex: string;
-    public description: string;
- }
