+++ src/gluon/events/team/MembershipRequestClosed.ts
@@ -14,7 +14,7 @@ import {
 } from "@atomist/automation-client/lib/spi/message/MessageClient";
 import {SlackMessage} from "@atomist/slack-messages";
 import {isSuccessCode} from "../../../http/Http";
-import {QMColours} from "../../../QMColour";
+import {QMColours} from "../../util/QMColour";
 import {GluonService} from "../../services/gluon/GluonService";
 import {BaseQMEvent} from "../../util/shared/BaseQMEvent";
 import {
