+++ src/gluon/services/gluon/ApplicationService.ts
@@ -5,7 +5,7 @@ import * as _ from "lodash";
 import {QMConfig} from "../../../config/QMConfig";
 import {AwaitAxios} from "../../../http/AwaitAxios";
 import {isSuccessCode} from "../../../http/Http";
-import {QMColours} from "../../../QMColour";
+import {QMColours} from "../../util/QMColour";
 import {LinkExistingApplication} from "../../commands/packages/LinkExistingApplication";
 import {QMError} from "../../util/shared/Error";
