+++ src/gluon/services/gluon/ProjectService.ts
@@ -5,7 +5,7 @@ import _ = require("lodash");
 import {QMConfig} from "../../../config/QMConfig";
 import {AwaitAxios} from "../../../http/AwaitAxios";
 import {isSuccessCode} from "../../../http/Http";
-import {QMColours} from "../../../QMColour";
+import {QMColours} from "../../util/QMColour";
 import {CreateProject} from "../../commands/project/CreateProject";
 import {QMError} from "../../util/shared/Error";
