+++ src/gluon/tasks/TaskListMessage.ts
@@ -1,7 +1,7 @@
 import {HandlerResult, logger} from "@atomist/automation-client";
 import {Attachment, SlackMessage} from "@atomist/slack-messages";
 import {v4 as uuid} from "uuid";
-import {QMColours} from "../../QMColour";
+import {QMColours} from "../util/QMColour";
 import {QMMessageClient} from "../util/shared/Error";
 
 export class TaskListMessage {
