+++ src/gluon/util/QMColour.ts
@@ -1,0 +1,30 @@
+export class QMColours {
+
+    public static absaAgile: QMColour = { name: "absaAgile", hex: "#f52d28", description: "absa Agile" };
+    public static absaBlack: QMColour = { name: "absaBlack", hex: "#000000", description: "absa Black" };
+    public static absaCalm: QMColour = { name: "absaCalm", hex: "#870a3c", description: "absa Calm" };
+    public static absaCare: QMColour = { name: "absaCare", hex: "#f05a7d", description: "absa Care" };
+    public static absaDepth: QMColour = { name: "absaDepth", hex: "#500a28", description: "absa Depth" };
+    public static absaEnergy: QMColour = { name: "absaEnergy", hex: "#ff780f", description: "absa Energy" };
+    public static absaGrounded: QMColour = { name: "absaGrounded", hex: "#960528", description: "absa Grounded" };
+    public static absaHuman: QMColour = { name: "absaHuman", hex: "#aa052d", description: "absa Human" };
+    public static absaLuxury: QMColour = { name: "absaLuxury", hex: "#640032", description: "absa Luxury" };
+    public static absaPassion: QMColour = { name: "absaPassion", hex: "#dc0032", description: "absa Passion" };
+    public static absaPrepared: QMColour = { name: "absaPrepared", hex: "#fa551e", description: "absa Prepared" };
+    public static absaSmile: QMColour = { name: "absaSmile", hex: "#f0325a", description: "absa Smile" };
+    public static absaSurprise: QMColour = { name: "absaSurprise", hex: "#af144b", description: "absaS urprise" };
+    public static absaWarmth: QMColour = { name: "absaWarmth", hex: "#be0028", description: "absa Warmth" };
+
+    public static stdGreenyMcAppleStroodle: QMColour = { name: "stdGreenyMcAppleStroodle", hex: "#45B254", description: "Success or default option" };
+    public static stdMuddyYellow: QMColour = { name: "stdMuddyYellow", hex: "#ffcc00", description: "Warning" };
+    public static stdPurplePeopleEater: QMColour = { name: "stdPurplePeopleEater", hex: "#c000ff", description: "A purple colour" };
+    public static stdReddyMcRedFace: QMColour = { name: "stdReddyMcRedFace", hex: "#D94649", description: "Error" };
+    public static stdShySkyBlue: QMColour = { name: "stdShySkyBlue", hex: "#00a5ff", description: "Secondary option or user prompt" };
+    public static stdTurquoiseSurprise: QMColour = { name: "stdTurquoise Surprise", hex: "#00ddff", description: "A turquoise colour" };
+ }
+
+export class QMColour {
+    public name: string;
+    public hex: string;
+    public description: string;
+ }
