+++ src/gluon/util/help/Help.ts
@@ -9,7 +9,7 @@ import {HandleCommand} from "@atomist/automation-client/lib/HandleCommand";
 import {buttonForCommand} from "@atomist/automation-client/lib/spi/message/MessageClient";
 import uuid = require("uuid");
 import {QMConfig} from "../../../config/QMConfig";
-import {QMColours} from "../../../QMColour";
+import {QMColours} from "../QMColour";
 import {handleQMError, ResponderMessageClient} from "../shared/Error";
 import {HelpCategory} from "./HelpCategory";
